Refactor: 스타일, 타입 이동 (#152)
