+++ modules/post/components/detail/style.ts
@@ -1,9 +1,5 @@
 import { makeStyles } from '@rneui/themed';
 
-export default makeStyles((theme) => ({
-  dividerStyle: {
-    width: 12,
-    color: theme.font.grayscale['100'],
-  },
+export default makeStyles(() => ({
   cContentHeader: { flexDirection: 'row', marginTop: 28, marginBottom: 12 },
 }));
