+++ modules/post/components/detail/type.ts
@@ -1,9 +1,5 @@
 import { ReactNode } from 'react';
 
-export interface IcontentHeader {
-  children: ReactNode;
-}
-
 export interface IdividerWrapper {
   children: ReactNode;
 }
