+++ modules/shared/components/layout/style.ts
@@ -1,6 +1,10 @@
 import { makeStyles } from '@rneui/themed';
 
-export default makeStyles(() => ({
+export default makeStyles((theme) => ({
+  dividerStyle: {
+    width: 12,
+    color: theme.font.grayscale['100'],
+  },
   container: {
     paddingVertical: 10,
     paddingHorizontal: 15,
