+++ modules/shared/components/layout/type.ts
@@ -1,6 +1,10 @@
 import { ReactNode } from 'react';
 import { StyleProp, ViewStyle } from 'react-native';
 
+export interface IcontentHeader {
+  children: ReactNode;
+}
+
 export interface IscreenLayout {
   children: ReactNode;
   testID?: string;
